requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
tqdm
tensorflow